+++ server/test/forms/questions/DateQuestionFormTest.java
@@ -1,44 +1,42 @@
-package forms;
+package forms.questions;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.assertThrows;
+import static services.question.types.DateQuestionDefinition.DateValidationOption.DateType.ANY;
+import static services.question.types.DateQuestionDefinition.DateValidationOption.DateType.APPLICATION_DATE;
 
-import forms.questions.FileUploadQuestionForm;
 import java.util.Locale;
-import java.util.OptionalInt;
+import java.util.Optional;
 import java.util.UUID;
 import org.junit.Test;
 import services.LocalizedStrings;
-import services.question.types.FileUploadQuestionDefinition;
+import services.question.types.DateQuestionDefinition;
+import services.question.types.DateQuestionDefinition.DateValidationOption;
+import services.question.types.DateQuestionDefinition.DateValidationPredicates;
 import services.question.types.QuestionDefinition;
 import services.question.types.QuestionDefinitionBuilder;
 import services.question.types.QuestionDefinitionConfig;
 
-public class FileUploadQuestionFormTest {
+public class DateQuestionFormTest {
+
   @Test
   public void getBuilder_returnsCompleteBuilder() throws Exception {
     UUID initialToken = UUID.randomUUID();
-    FileUploadQuestionForm form = new FileUploadQuestionForm();
-    form.setQuestionName("file upload");
+    DateQuestionForm form = new DateQuestionForm();
+    form.setQuestionName("name");
     form.setQuestionDescription("description");
     form.setQuestionText("What is the question text?");
     form.setQuestionHelpText("");
     form.setConcurrencyToken(initialToken);
-    form.setMaxFiles("4");
     QuestionDefinitionBuilder builder = form.getBuilder();
 
-    FileUploadQuestionDefinition expected =
-        new FileUploadQuestionDefinition(
+    DateQuestionDefinition expected =
+        new DateQuestionDefinition(
             QuestionDefinitionConfig.builder()
-                .setName("file upload")
+                .setName("name")
                 .setDescription("description")
                 .setQuestionText(LocalizedStrings.of(Locale.US, "What is the question text?"))
                 .setQuestionHelpText(LocalizedStrings.empty())
                 .setConcurrencyToken(initialToken)
-                .setValidationPredicates(
-                    FileUploadQuestionDefinition.FileUploadValidationPredicates.builder()
-                        .setMaxFiles(OptionalInt.of(4))
-                        .build())
                 .build());
 
     QuestionDefinition actual = builder.build();
@@ -48,44 +46,27 @@ public class FileUploadQuestionFormTest {
 
   @Test
   public void getBuilder_withQdConstructor_returnsCompleteBuilder() throws Exception {
-    FileUploadQuestionDefinition originalQd =
-        new FileUploadQuestionDefinition(
+    DateValidationOption minDate =
+        DateValidationOption.builder().setDateType(APPLICATION_DATE).build();
+    DateValidationOption maxDate = DateValidationOption.builder().setDateType(ANY).build();
+    DateValidationPredicates validationPredicates =
+        DateValidationPredicates.create(Optional.of(minDate), Optional.of(maxDate));
+    DateQuestionDefinition originalQd =
+        new DateQuestionDefinition(
             QuestionDefinitionConfig.builder()
-                .setName("file upload")
+                .setName("name")
                 .setDescription("description")
                 .setQuestionText(LocalizedStrings.of(Locale.US, "What is the question text?"))
                 .setQuestionHelpText(LocalizedStrings.empty())
                 .setConcurrencyToken(UUID.randomUUID())
-                .setValidationPredicates(
-                    FileUploadQuestionDefinition.FileUploadValidationPredicates.builder()
-                        .setMaxFiles(OptionalInt.of(4))
-                        .build())
+                .setValidationPredicates(validationPredicates)
                 .build());
 
-    FileUploadQuestionForm form = new FileUploadQuestionForm(originalQd);
+    DateQuestionForm form = new DateQuestionForm(originalQd);
     QuestionDefinitionBuilder builder = form.getBuilder();
 
     QuestionDefinition actual = builder.build();
 
     assertThat(actual).isEqualTo(originalQd);
-  }
-
-  @Test
-  public void setMaxFiles_emptyStringClearsMaxFiles() throws Exception {
-    FileUploadQuestionForm form = new FileUploadQuestionForm();
-    form.setMaxFiles("4");
-
-    assertThat(form.getMaxFiles().getAsInt()).isEqualTo(4);
-
-    form.setMaxFiles("");
-
-    assertThat(form.getMaxFiles()).isEmpty();
-  }
-
-  @Test
-  public void setMaxFiles_invalidStringThrowsException() throws Exception {
-    FileUploadQuestionForm form = new FileUploadQuestionForm();
-
-    assertThrows(NumberFormatException.class, () -> form.setMaxFiles("four"));
   }
 }
