+++ server/test/forms/questions/EnumeratorQuestionFormTest.java
@@ -1,8 +1,7 @@
-package forms;
+package forms.questions;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import forms.questions.EnumeratorQuestionForm;
 import java.util.Locale;
 import java.util.OptionalInt;
 import java.util.UUID;
