+++ server/test/forms/questions/PhoneQuestionFormTest.java
@@ -1,36 +1,34 @@
-package forms;
+package forms.questions;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import forms.questions.CurrencyQuestionForm;
 import java.util.Locale;
 import java.util.UUID;
 import org.junit.Test;
 import services.LocalizedStrings;
-import services.question.types.CurrencyQuestionDefinition;
+import services.question.types.PhoneQuestionDefinition;
 import services.question.types.QuestionDefinition;
 import services.question.types.QuestionDefinitionBuilder;
 import services.question.types.QuestionDefinitionConfig;
 
-public class CurrencyQuestionFormTest {
-
+public class PhoneQuestionFormTest {
   @Test
   public void getBuilder_returnsCompleteBuilder() throws Exception {
     UUID initialToken = UUID.randomUUID();
-    CurrencyQuestionForm form = new CurrencyQuestionForm();
+    PhoneQuestionForm form = new PhoneQuestionForm();
     form.setQuestionName("name");
     form.setQuestionDescription("description");
-    form.setQuestionText("What is the question text?");
+    form.setQuestionText("What is your phone number?");
     form.setQuestionHelpText("");
     form.setConcurrencyToken(initialToken);
     QuestionDefinitionBuilder builder = form.getBuilder();
 
-    CurrencyQuestionDefinition expected =
-        new CurrencyQuestionDefinition(
+    PhoneQuestionDefinition expected =
+        new PhoneQuestionDefinition(
             QuestionDefinitionConfig.builder()
                 .setName("name")
                 .setDescription("description")
-                .setQuestionText(LocalizedStrings.of(Locale.US, "What is the question text?"))
+                .setQuestionText(LocalizedStrings.of(Locale.US, "What is your phone number?"))
                 .setQuestionHelpText(LocalizedStrings.empty())
                 .setConcurrencyToken(initialToken)
                 .build());
@@ -42,17 +40,17 @@ public class CurrencyQuestionFormTest {
 
   @Test
   public void getBuilder_withQdConstructor_returnsCompleteBuilder() throws Exception {
-    CurrencyQuestionDefinition originalQd =
-        new CurrencyQuestionDefinition(
+    PhoneQuestionDefinition originalQd =
+        new PhoneQuestionDefinition(
             QuestionDefinitionConfig.builder()
                 .setName("name")
                 .setDescription("description")
-                .setQuestionText(LocalizedStrings.of(Locale.US, "What is the question text?"))
+                .setQuestionText(LocalizedStrings.of(Locale.US, "What is your Phone Number?"))
                 .setQuestionHelpText(LocalizedStrings.empty())
                 .setConcurrencyToken(UUID.randomUUID())
                 .build());
 
-    CurrencyQuestionForm form = new CurrencyQuestionForm(originalQd);
+    PhoneQuestionForm form = new PhoneQuestionForm(originalQd);
     QuestionDefinitionBuilder builder = form.getBuilder();
 
     QuestionDefinition actual = builder.build();
