+++ server/test/forms/questions/QuestionFormTest.java
@@ -1,10 +1,9 @@
-package forms;
+package forms.questions;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.google.common.collect.ImmutableSet;
-import forms.questions.QuestionForm;
 import java.util.UUID;
 import models.QuestionDisplayMode;
 import org.junit.Test;
